ORC-176: Refactor common classes for writer and reader
This is mainly a refactoring change for ORC-176, including:

1. Extracted common classes and functions into Common.hh and Common.cc;
2. Put InputStream interface and its implementations from Compression.hh to InputStream.hh and InputStream.cc.
