+++ c++/include/orc/Reader.hh
@@ -19,6 +19,7 @@
 #ifndef ORC_READER_HH
 #define ORC_READER_HH
 
+#include "orc/Common.hh"
 #include "orc/orc-config.hh"
 #include "orc/Statistics.hh"
 #include "orc/Type.hh"
@@ -33,142 +34,6 @@ namespace orc {
   // classes that hold data members so we can maintain binary compatibility
   struct ReaderOptionsPrivate;
   struct RowReaderOptionsPrivate;
-
-  enum CompressionKind {
-    CompressionKind_NONE = 0,
-    CompressionKind_ZLIB = 1,
-    CompressionKind_SNAPPY = 2,
-    CompressionKind_LZO = 3,
-    CompressionKind_LZ4 = 4,
-    CompressionKind_ZSTD = 5,
-    CompressionKind_MAX = INT64_MAX
-  };
-
-  /**
-   * Get the name of the CompressionKind.
-   */
-  std::string compressionKindToString(CompressionKind kind);
-
-  enum WriterVersion {
-    WriterVersion_ORIGINAL = 0,
-    WriterVersion_HIVE_8732 = 1,
-    WriterVersion_HIVE_4243 = 2,
-    WriterVersion_HIVE_12055 = 3,
-    WriterVersion_HIVE_13083 = 4,
-    WriterVersion_ORC_101 = 5,
-    WriterVersion_ORC_135 = 6,
-    WriterVersion_MAX = INT64_MAX
-  };
-
-  /**
-   * Get the name of the WriterVersion.
-   */
-  std::string writerVersionToString(WriterVersion kind);
-
-  enum StreamKind {
-    StreamKind_PRESENT = 0,
-    StreamKind_DATA = 1,
-    StreamKind_LENGTH = 2,
-    StreamKind_DICTIONARY_DATA = 3,
-    StreamKind_DICTIONARY_COUNT = 4,
-    StreamKind_SECONDARY = 5,
-    StreamKind_ROW_INDEX = 6,
-    StreamKind_BLOOM_FILTER = 7
-  };
-
-  /**
-   * Get the string representation of the StreamKind.
-   */
-  std::string streamKindToString(StreamKind kind);
-
-  class StreamInformation {
-  public:
-    virtual ~StreamInformation();
-
-    virtual StreamKind getKind() const = 0;
-    virtual uint64_t getColumnId() const = 0;
-    virtual uint64_t getOffset() const = 0;
-    virtual uint64_t getLength() const = 0;
-  };
-
-  enum ColumnEncodingKind {
-    ColumnEncodingKind_DIRECT = 0,
-    ColumnEncodingKind_DICTIONARY = 1,
-    ColumnEncodingKind_DIRECT_V2 = 2,
-    ColumnEncodingKind_DICTIONARY_V2 = 3
-  };
-
-  std::string columnEncodingKindToString(ColumnEncodingKind kind);
-
-  class StripeInformation {
-  public:
-    virtual ~StripeInformation();
-
-    /**
-     * Get the byte offset of the start of the stripe.
-     * @return the bytes from the start of the file
-     */
-    virtual uint64_t getOffset() const = 0;
-
-    /**
-     * Get the total length of the stripe in bytes.
-     * @return the number of bytes in the stripe
-     */
-    virtual uint64_t getLength() const = 0;
-
-    /**
-     * Get the length of the stripe's indexes.
-     * @return the number of bytes in the index
-     */
-    virtual uint64_t getIndexLength() const = 0;
-
-    /**
-     * Get the length of the stripe's data.
-     * @return the number of bytes in the stripe
-     */
-    virtual uint64_t getDataLength()const = 0;
-
-    /**
-     * Get the length of the stripe's tail section, which contains its index.
-     * @return the number of bytes in the tail
-     */
-    virtual uint64_t getFooterLength() const = 0;
-
-    /**
-     * Get the number of rows in the stripe.
-     * @return a count of the number of rows
-     */
-    virtual uint64_t getNumberOfRows() const = 0;
-
-    /**
-     * Get the number of streams in the stripe.
-     */
-    virtual uint64_t getNumberOfStreams() const = 0;
-
-    /**
-     * Get the StreamInformation for the given stream.
-     */
-    virtual ORC_UNIQUE_PTR<StreamInformation>
-      getStreamInformation(uint64_t streamId) const = 0;
-
-    /**
-     * Get the column encoding for the given column.
-     * @param colId the columnId
-     */
-    virtual ColumnEncodingKind getColumnEncoding(uint64_t colId) const = 0;
-
-    /**
-     * Get the dictionary size.
-     * @param colId the columnId
-     * @return the size of the dictionary or 0 if there isn't one
-     */
-    virtual uint64_t getDictionarySize(uint64_t colId) const = 0;
-
-    /**
-     * Get the writer timezone.
-     */
-    virtual const std::string& getWriterTimezone() const = 0;
-  };
 
   /**
    * Options for creating a Reader.
